--- StonksMobile/app/api/Endpoints.js
@@ -1,9 +1,0 @@
-const endpoints = {
-    getStock: "/stock",
-    getStockBySymbol: (symbol) => `/stock/${symbol}`,
-
-    getQoute:(symbol) => `/api/Stocks/${symbol}/quote`
-    
-}
-
-export default endpoints;
