--- StonksMobile/app/screens/createAccountPage/CreateAccountStyles.js
@@ -1,5 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const CreateAccountStyles = StyleSheet.create({
-    text: {}
-})
