--- StonksMobile/app/screens/loginScreen/LoginStyles.js
@@ -1,17 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const LoginStyles = StyleSheet.create({
-    componentContainer : {
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-    },
-    infoContainer : {
-        flexShrink: 1,
-        marginBottom: 20,
-        justifyContent: "flex-end",
-        alignItems: "center",
-    }
-});
-
-export default LoginStyles;
