+++ StonksMobile/src/assets/rneIcons/FeatherIcon.js
