+++ StonksMobile/src/assets/rneIcons/Ionicons.js
