+++ StonksMobile/src/assets/test-chart1.png
