+++ StonksMobile/src/features/home-overview/store/MoverConsts.js
