+++ StonksMobile/src/features/market-info/components/StatsBox.js
