+++ StonksMobile/src/libs/Endpoints.js
@@ -1,0 +1,8 @@
+const endpoints = {
+    getStock: "/stock",
+    getStockBySymbol: (symbol) => `/stock/${symbol}`,
+    getQoute:(symbol) => `/api/Stocks/${symbol}/quote`
+    
+}
+
+export default endpoints;
