+++ StonksMobile/src/theme/AppStyles.js
